+++ plugin/plugin_destination.go
@@ -2,6 +2,7 @@ package plugin
 
 import (
 	"context"
+	"errors"
 	"fmt"
 
 	"github.com/apache/arrow-go/v18/arrow"
@@ -33,7 +34,7 @@ func (p *Plugin) WriteAll(ctx context.Context, resources []message.WriteMessage)
 
 func (p *Plugin) Write(ctx context.Context, res <-chan message.WriteMessage) error {
 	if p.client == nil {
-		return fmt.Errorf("plugin is not initialized. call Init first")
+		return errors.New("plugin is not initialized. call Init first")
 	}
 	return p.client.Write(ctx, res)
 }
@@ -41,11 +42,11 @@ func (p *Plugin) Write(ctx context.Context, res <-chan message.WriteMessage) err
 // Read is read data from the requested table to the given channel, returned in the same format as the table
 func (p *Plugin) Read(ctx context.Context, table *schema.Table, res chan<- arrow.Record) error {
 	if !p.mu.TryLock() {
-		return fmt.Errorf("plugin already in use")
+		return errors.New("plugin already in use")
 	}
 	defer p.mu.Unlock()
 	if p.client == nil {
-		return fmt.Errorf("plugin not initialized. call Init() first")
+		return errors.New("plugin not initialized. call Init() first")
 	}
 	if err := p.client.Read(ctx, table, res); err != nil {
 		return fmt.Errorf("failed to read: %w", err)